requirements.txt
alembic==1.10.2
Babel==2.12.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
Flask==2.2.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-translate==3.11.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
